+++ packages/legend-application-query/src/components/QueryBuilderProjectionPanel.tsx
@@ -28,6 +28,7 @@ import {
   ContextMenu,
   InputWithInlineValidation,
   SigmaIcon,
+  TimesCircleIcon,
 } from '@finos/legend-art';
 import {
   type QueryBuilderExplorerTreeDragSource,
@@ -37,7 +38,6 @@ import {
 } from '../stores/QueryBuilderExplorerState.js';
 import {
   type DropTargetMonitor,
-  type XYCoord,
   useDragLayer,
   useDrag,
   useDrop,
@@ -341,8 +341,7 @@ const QueryBuilderProjectionColumnEditor = observer(
           ((hoverBoundingReact?.bottom ?? 0) - (hoverBoundingReact?.top ?? 0)) /
           2;
         const dragDistance =
-          (monitor.getClientOffset() as XYCoord).y -
-          (hoverBoundingReact?.top ?? 0);
+          (monitor.getClientOffset()?.y ?? 0) - (hoverBoundingReact?.top ?? 0);
         if (dragIndex < hoverIndex && dragDistance < distanceThreshold) {
           return;
         }
@@ -588,6 +587,8 @@ export const QueryBuilderProjectionPanel = observer(
       },
       [queryBuilderState, projectionState],
     );
+
+    const isInvalidProjection = !projectionState.isValidProjectionState();
     const [{ isPropertyDragOver }, dropConnector] = useDrop(
       () => ({
         accept: [
@@ -620,10 +621,21 @@ export const QueryBuilderProjectionPanel = observer(
 
     return (
       <div
-        className="panel__content dnd__overlay__container"
         ref={dropConnector}
+        className={`panel__content dnd__overlay__container`}
       >
+        {isInvalidProjection && (
+          <div className="query-builder__projection__container__error">
+            <div className="query-builder__projection__error__label">
+              <TimesCircleIcon className="query-builder__projection__error__icon" />
+              {projectionState.getValidationErrorMessage() ??
+                'There is an error with the projection'}
+            </div>
+          </div>
+        )}
+
         <div className={clsx({ dnd__overlay: isPropertyDragOver })} />
+
         {!projectionColumns.length && (
           <BlankPanelPlaceholder
             placeholderText="Add a projection column"
@@ -633,7 +645,10 @@ export const QueryBuilderProjectionPanel = observer(
         {Boolean(projectionColumns.length) && (
           <div
             data-testid={QUERY_BUILDER_TEST_ID.QUERY_BUILDER_PROJECTION}
-            className="query-builder__projection__columns"
+            className={`query-builder__projection__columns ${clsx({
+              ['query-builder__projection__columns--with-error']:
+                isInvalidProjection,
+            })} `}
           >
             <ProjectionColumnDragLayer />
             {projectionColumns.map((projectionColumnState) => (