+++ packages/legend-extension-dsl-mastery/src/graph/DSLMastery_HashUtils.ts
@@ -1,0 +1,23 @@
+/**
+ * Copyright (c) 2020-present, Goldman Sachs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+export enum MASTERY_HASH_STRUCTURE {
+  MASTER_RECORD_DEFINITION = 'MASTER_RECORD_DEFINITION',
+  IDENTITY_RESOLUTION = 'IDENTITY_RESOLUTION',
+  RESOLUTION_QUERY = 'RESOLUTION_QUERY',
+  RECORD_SOURCE = 'RECORD_SOURCE',
+  RECORD_SOURCE_PARTITION = 'RECORD_SOURCE_PARTITION',
+}