+++ packages/legend-application-taxonomy-bootstrap/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @finos/legend-application-taxonomy-bootstrap
 
+## 7.3.0
+
+## 7.2.0
+
 ## 7.1.0
 
 ## 7.0.0