+++ packages/legend-application-taxonomy/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @finos/legend-application-taxonomy
 
+## 6.0.3
+
+## 6.0.2
+
 ## 6.0.1
 
 ## 6.0.0