+++ packages/legend-application/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @finos/legend-application
 
+## 7.0.4
+
+### Patch Changes
+
+- [#1381](https://github.com/finos/legend-studio/pull/1381) [`7a892e6b`](https://github.com/finos/legend-studio/commit/7a892e6b3906b7429ee89ee0fe573e4215553fbf) ([@MauricioUyaguari](https://github.com/MauricioUyaguari)) - Add support for selection of value specification values through options.
+
+## 7.0.3
+
+### Patch Changes
+
+- [#1378](https://github.com/finos/legend-studio/pull/1378) [`e6812be0`](https://github.com/finos/legend-studio/commit/e6812be09157cbceb52744c55533d5fc094cb119) ([@MauricioUyaguari](https://github.com/MauricioUyaguari)) - Fix generating expected results for multi execution ([#1352](https://github.com/finos/legend-studio/issues/1352)).
+
 ## 7.0.2
 
 ## 7.0.1