+++ packages/legend-extension-dsl-data-space/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @finos/legend-extension-dsl-data-space
 
+## 6.1.2
+
+## 6.1.1
+
 ## 6.1.0
 
 ### Minor Changes