+++ packages/legend-extension-dsl-text/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @finos/legend-extension-dsl-text
 
+## 4.0.3
+
+## 4.0.2
+
 ## 4.0.1
 
 ## 4.0.0