+++ packages/legend-extension-external-format-json-schema/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @finos/legend-extension-external-format-json-schema
 
+## 3.0.3
+
+## 3.0.2
+
 ## 3.0.1
 
 ## 3.0.0