+++ packages/legend-graph/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @finos/legend-graph
 
+## 13.0.2
+
+## 13.0.1
+
 ## 13.0.0
 
 ### Major Changes