+++ packages/legend-server-depot/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @finos/legend-server-depot
 
+## 3.0.4
+
+## 3.0.3
+
 ## 3.0.2
 
 ## 3.0.1